api_service/requirements.txt
Flask==1.0.2
gunicorn==19.8.1
requests==2.19.1
google-cloud-pubsub==0.40.0